@@ -1,13 +1,13 @@
 name: picpay
 description: PicPay E-Commerce API for Dart and Flutter for create Payments and check Payment status
-version: 1.1.1
+version: 1.1.2
 homepage: https://github.com/stackingwidgets/picpay-dart
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
-  http: ^0.12.2
+  http: ^0.13.4
 
 dev_dependencies:
   pedantic: any
